display item image and desc separately
